powf is microsoft

+++ src/Eikonal.cu
@@ -1,4 +1,5 @@
 #include "Eikonal.h"
+#include <cmath>
 
 Eikonal::Eikonal(bool isTriMesh, std::string fname, bool verbose) :
   verbose_(verbose),
@@ -246,9 +247,9 @@ void Eikonal::printErrorGraph(std::vector<float> solution) {
   float max_err = rmsError[0];
   float min_err = rmsError[rmsError.size() - 1];
   int max_log = -10, min_log = 10;
-  while (std::powf(static_cast<float>(10),
+  while (std::pow(static_cast<float>(10),
         static_cast<float>(max_log)) < max_err) max_log++;
-  while (std::powf(static_cast<float>(10),
+  while (std::pow(static_cast<float>(10),
         static_cast<float>(min_log)) > min_err) min_log--;
   // print the error graph
   printf("\n\nlog(Err)|\n");
@@ -260,9 +261,9 @@ void Eikonal::printErrorGraph(std::vector<float> solution) {
       printf("        |");
     }
     for (size_t j = 0; j < numIterations(); j++) {
-      if (rmsError[j] > std::powf(static_cast<float>(10),
+      if (rmsError[j] > std::pow(static_cast<float>(10),
             static_cast<float>(i)) &&
-          rmsError[j] < std::powf(static_cast<float>(10),
+          rmsError[j] < std::pow(static_cast<float>(10),
             static_cast<float>(i + 1)))
         printf("*");
       else
